+++ resolwe/flow/models/data.py
@@ -634,6 +634,33 @@ class Data(HistoryMixin, PermissionObject, BaseModel):
                 self.tags = entity.tags
         self.save(update_fields=["permission_group", "entity", "tags"])
 
+    @classmethod
+    def validate_change_containers(cls, data, entity, collection):
+        """Validate changing entity and collection.
+
+        Validate that:
+        - data is not removed from the container.
+        - if entity is given its collection must be the same as given collection
+        """
+        if data and data.in_container():
+            if (data.entity and not entity) or (data.collection and not collection):
+                raise ValidationError(
+                    "Data object can not be removed from the container."
+                )
+        if entity and entity.collection != collection:
+            raise ValidationError(
+                "Entity must belong to the same collection as data object."
+            )
+
+    def move_to_containers(self, entity, collection):
+        """Move the data object to the given entity and collection."""
+        Data.validate_change_containers(self, entity, collection)
+        if self.entity != entity or self.collection != collection:
+            self.collection = collection
+            self.entity = entity
+            self.tags = getattr(collection, "tags", []) or getattr(entity, "tags", [])
+            self.save(update_fields=["collection", "entity", "tags"])
+
     def validate_change_collection(self, collection):
         """Raise validation error if data object cannot change collection."""
         if self.entity and self.entity.collection != collection:
@@ -643,6 +670,17 @@ class Data(HistoryMixin, PermissionObject, BaseModel):
             )
         if collection is None:
             raise ValidationError("Data object can not be removed from the container.")
+
+    def validate_change_entity(self, entity):
+        """Raise validation error if data object cannot change entity."""
+        if entity is None:
+            raise ValidationError("Data object can not be removed from the container.")
+
+        if self.collection and self.collection != entity.collection:
+            raise ValidationError(
+                "If data is in collection, you can only move it to another entity "
+                "in the same collection."
+            )
 
     def _render_name(self):
         """Render data name.